Move Chat and Channels services to ChatView
