+++ src/app/chat/chat-view/channels.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
 import { Apollo, ApolloQueryObservable } from 'apollo-angular';
 
-import { ChannelByNameQuery, DirectChannelQuery, MyChannelsQuery } from '../../../graphql/types/types';
-import { AuthenticationService } from '../../../shared/services/authentication.service';
-import { myChannelsQuery } from '../../../graphql/queries/my-channels.query';
-import { channelByNameQuery } from '../../../graphql/queries/channel-by-name.query';
-import { directChannelQuery } from '../../../graphql/queries/direct-channel.query';
+import { ChannelByNameQuery, DirectChannelQuery, MyChannelsQuery } from '../../graphql/types/types';
+import { AuthenticationService } from '../../shared/services/authentication.service';
+import { myChannelsQuery } from '../../graphql/queries/my-channels.query';
+import { channelByNameQuery } from '../../graphql/queries/channel-by-name.query';
+import { directChannelQuery } from '../../graphql/queries/direct-channel.query';
 
 @Injectable()
 export class ChannelsService {
