+++ src/app/chat/chat.module.ts
@@ -8,9 +8,9 @@ import { SharedModule } from '../shared/shared.module';
 import { MainPageComponent } from './main-page/main-page.component';
 import { ChannelItemComponent } from './channel-item/channel-item.component';
 import { ChatViewComponent } from './chat-view/chat-view.component';
-import { ChatService } from './services/chat/chat.service';
+import { ChatService } from './chat-view/chat.service';
 import { MainSidenavComponent } from './main-sidenav/main-sidenav.component';
-import { ChannelsService } from './services/channels/channels.service';
+import { ChannelsService } from './chat-view/channels.service';
 import { ChannelNotFoundComponent } from './channel-not-found/channel-not-found.component';
 import { TranslateServerPipe } from './pipes/translate-server.pipe';
