+++ src/app/chat/main-sidenav/main-sidenav.component.ts
@@ -5,7 +5,7 @@ import { MenuController } from 'ionic-angular';
 
 import { AuthenticationService } from '../../shared/services/authentication.service';
 import { UserFields } from '../../graphql/types/types';
-import { ChannelsService } from '../services/channels/channels.service';
+import { ChannelsService } from '../chat-view/channels.service';
 
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/do';
